client/requirements.txt
aiohttp==3.6.2
pyserial==3.5
